requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
bottle==0.13.4
Brotli==1.1.0
browser-cookie3==0.20.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clr_loader==0.2.7.post0
frozenlist==1.7.0
g4f==0.5.6.3
idna==3.10
lz4==4.4.4
MouseInfo==0.1.3
multidict==6.6.3
nest-asyncio==1.6.0
packaging==25.0
pillow==11.2.1
pynput==1.7.6
propcache==0.3.2
proxy-tools==0.1.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
PyGetWindow==0.0.9
pyinstaller==6.12.0
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pythonnet==3.0.5
pytweening==1.2.0
pywebview==5.4
pywin32==310
requests==2.32.4
selenium==4.23.1
shadowcopy==0.0.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
WMI==1.5.1
yarl==1.20.1
